requirements.txt
python-dotenv==1.2.1
django==6.0.1
pillow<=12.1.0
django-tables2==2.8.0
django-crispy-forms==2.5
crispy-daisyui==0.9.0
celery==5.6.2
pyexiftool==0.5.6
djangorestframework==3.16.1
redis==7.1.0
mozilla-django-oidc==5.0.2
psycopg2-binary ==2.9.11
gunicorn==24.1.1
django-celery-results==2.6.0
drf-spectacular==0.29.0
django-filter==25.2